after essai or abonnement version

+++ database/migrations/2026_09_11_000020_create_devis_table.php
@@ -9,7 +9,7 @@ return new class extends Migration {
     {
         Schema::create('devis', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('entreprise_id');
+            $table->unsignedBigInteger('entreprise_id');
             $table->unsignedBigInteger('user_id')->nullable();
             $table->string('numero_devis');
             $table->unsignedInteger('num');
