+++ database/migrations/2026_09_11_000051_create_reglement_clients_table.php
@@ -9,7 +9,7 @@ return new class extends Migration {
     {
         Schema::create('reglement_clients', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('entreprise_id');
+            $table->unsignedBigInteger('entreprise_id');
             $table->unsignedBigInteger('user_id')->nullable();
             $table->string('numero')->unique();
             $table->foreignId('client_id')->constrained('clients')->restrictOnDelete();
